requirements.txt
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.2.0
torchvision>=0.17.0
faiss-cpu>=1.8.0
pillow>=10.0.0
pandas>=2.0.0
requests>=2.31.0
streamlit>=1.35.0
accelerate>=0.29.0
numpy>=1.26.0
